Move/expose the IRuntimeContext #166
